requirements.txt
fastapi
uvicorn
motor
python-decouple
python-dotenv